requirements.txt
uvicorn[standard]
pydantic[email]
fastapi
deta
python-dotenv
pyjwt
python-decouple